Split codex turn logic
